+++ src/engine/Actor.ts
@@ -1336,12 +1336,14 @@ export class ActorImpl extends Class implements Actionable, Eventable, PointerEv
    * @param group The group name to listen for
    * @param func The callback to fire on collision with another actor from the group. The callback is passed the other actor.
    */
+  @obsolete({ message: 'Actor.onCollidesWIth will be removed  in v0.24.0', alternateMethod: 'Actor.collider.canCollide' })
   public onCollidesWith(group: string, func: (actor: Actor) => void) {
     if (!this._collisionHandlers[group]) {
       this._collisionHandlers[group] = [];
     }
     this._collisionHandlers[group].push(func);
   }
+  @obsolete({ message: 'Actor.getCollisionHandlers will be removed  in v0.24.0' })
   public getCollisionHandlers(): { [key: string]: { (actor: Actor): void }[] } {
     return this._collisionHandlers;
   }
@@ -1349,6 +1351,7 @@ export class ActorImpl extends Class implements Actionable, Eventable, PointerEv
    * Removes all collision handlers for this group on this actor
    * @param group Group to remove all handlers for on this actor.
    */
+  @obsolete({ message: 'Actor.getCollisionHandlers will be removed  in v0.24.0' })
   public removeCollidesWith(group: string) {
     this._collisionHandlers[group] = [];
   }